requirements.txt
astroid==2.2.5
colorama==0.4.1
EasyProcess==0.2.7
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pillow==6.1.0
pip==19.0.3
pylint==2.3.1
pyscreenshot==0.5.1
pywin32==224
setuptools==40.8.0
six==1.12.0
typed-ast==1.4.0
wrapt==1.11.2
